+++ examples/rma_functions.hpp
@@ -8,7 +8,7 @@ namespace rmafunctions {
     struct RmaFunctionConfig {
         char client_ip_address[IPV4_ADDRESS_STRING_LENGTH];
         unsigned int client_port;
-        unsigned int rma_memory_in_bytes;
+        int64_t rma_memory_in_bytes;
     };
     
 }
